pyproject.toml
[project]
name = "nyaya-dhwani"
version = "0.2.0"
description = "Legal RAG ingestion and multilingual assistant (Databricks + Sarvam) with DPDP Act compliance features"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  # Align with Databricks Apps template (stable with Gradio 4.44 + huggingface_hub).
  "pandas~=2.2.3",
  "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
# Core vector index (tests + retrieval); sentence-transformers is separate (heavy).
rag = [
  # faiss-cpu 1.8+ uses numpy._core (NumPy 2+), which conflicts with databricks-connect (numpy<2).
  # faiss-cpu 1.7.x uses numpy.core and works with numpy 1.26 + Spark client libraries.
  "numpy>=1.24,<2",
  "faiss-cpu>=1.7.0,<1.8",
  "pyarrow>=14.0.0",
]
rag_embed = [
  "sentence-transformers>=2.2.0",
]
# Playground Get code parity (OpenAI SDK + Databricks AI Gateway base_url …/mlflow/v1).
llm_openai = [
  "openai>=1.30.0",
]
# Dharma-Guard intent firewall
firewall = [
  "scikit-learn>=1.3.0",
]
# Chitragupta steganography for data portability
steganography = [
  "Pillow>=10.0.0",
  "cryptography>=41.0.0",
]
# All new features combined
features = [
  "scikit-learn>=1.3.0",
  "Pillow>=10.0.0",
  "cryptography>=41.0.0",
]
# Gradio UI for Databricks Apps (`python app/main.py`).
# Pin huggingface_hub with Gradio 4.44 — matches Databricks template; avoids HfFolder drift when sentence-transformers upgrades hub.
app = [
  "gradio~=4.44.0",
  # Must match gradio 4.44.x wheel metadata (gradio-client==1.3.0); newer client breaks api_info JSON schema parsing.
  "gradio-client==1.3.0",
  "huggingface-hub~=0.35.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
  "network: tests that call external HTTP (opt-in)",
]
